ReinforcementLearning/requirements.txt
# Core dependencies
numpy>=1.25.0
pandas>=1.5.3
scikit-learn>=1.2.2
scipy>=1.10.1
pyyaml>=6.0

# Visualization
matplotlib>=3.8.3

# API
Flask>=3.0.3
Werkzeug>=3.0.1

# Testing
pytest>=7.4.0 